+++ bot/src/main/java/edu/java/bot/Bot.java
@@ -16,9 +16,9 @@ import edu.java.bot.processor.UpdateProcessor;
 import edu.java.bot.repository.UserRepository;
 import edu.java.bot.repository.UserRepositoryImpl;
 import edu.java.bot.service.TrackedLinkService;
-import edu.java.bot.service.TrackedLinkServiceImpl;
+import edu.java.bot.service.impl.TrackedLinkServiceImpl;
 import edu.java.bot.service.UserService;
-import edu.java.bot.service.UserServiceImpl;
+import edu.java.bot.service.impl.UserServiceImpl;
 import edu.java.bot.source.GitHubTrackableLink;
 import edu.java.bot.source.StackOverflowTrackableLink;
 import edu.java.bot.source.TrackableLink;
